requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
sqlmodel>=0.0.16
psycopg2-binary>=2.9.9
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.9
pillow>=10.4.0
python-decouple>=3.8
alembic>=1.13.1
pytest>=8.0.0
httpx>=0.27.0
pydantic-settings>=2.0.0